Allocation the translation module. Fix hover translation 403
