+++ translator/google-api.js
@@ -1,0 +1,3 @@
+exports.translate = function () {
+  throw new Error("Google Translate API isn't supported yet!");
+};
