+++ translator/google-translate-open-api.js
@@ -1,0 +1,3 @@
+exports.translate = function () {
+  throw new Error("The module google-translate-open-api isn't supported temporarily!");
+};
