+++ translator/test.js
@@ -1,0 +1,5 @@
+const t = require("./");
+
+(async () => {
+  console.log(await t.translate("test", { to: "ru" }));
+})();
